requirements.txt
ticktick-py
mdutils